+++ core/src/main/kotlin/com/bumble/appyx/core/minimal/reactive/Emitter.kt
@@ -1,5 +1,0 @@
-package com.bumble.appyx.core.minimal.reactive
-
-interface Emitter<in T> {
-    fun emit(value: T)
-}