+++ core/src/main/kotlin/com/bumble/appyx/core/minimal/reactive/Source.kt
@@ -1,5 +1,0 @@
-package com.bumble.appyx.core.minimal.reactive
-
-interface Source<out T> {
-    fun observe(callback: (T) -> Unit): Cancellable
-}