+++ samples/app/src/main/kotlin/com/bumble/appyx/app/MainActivity.kt
@@ -1,0 +1,55 @@
+package com.bumble.appyx.app
+
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.foundation.layout.Column
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.ExperimentalUnitApi
+import com.bumble.appyx.app.node.root.RootNode
+import com.bumble.appyx.app.ui.AppyxSampleAppTheme
+import com.bumble.appyx.core.integration.NodeHost
+import com.bumble.appyx.core.integrationpoint.NodeActivity
+import com.bumble.appyx.core.modality.BuildContext
+
+@ExperimentalUnitApi
+@ExperimentalAnimationApi
+@ExperimentalComposeUiApi
+class MainActivity : NodeActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            AppyxSampleAppTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(color = MaterialTheme.colors.background) {
+                    Column {
+                        NodeHost(integrationPoint = appyxIntegrationPoint) {
+                            RootNode(
+                                buildContext = it,
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+
+@Preview(showBackground = true)
+@Composable
+@ExperimentalUnitApi
+@ExperimentalAnimationApi
+@ExperimentalComposeUiApi
+fun DefaultPreview() {
+    AppyxSampleAppTheme {
+        Column {
+            RootNode(buildContext = BuildContext.root(null)).Compose()
+        }
+    }
+}